Move the worker output reader into its own file
TransformWorkerHost.cs reached 503 SLOC against the repository's 500 limit.
TransformWorkerOutputReader is the one type in that file with no tie to the
host's state machine, so it moves out unchanged.
