+++ Packages/src/Editor/FirstPartyTools/HotReload/TransformWorkerOutputReader.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f2ba772be517844d38c985e66c33fa9f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
